requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
django-cors-headers==3.14.0
django-filter==23.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gcloud==0.18.3
googleapis-common-protos==1.57.0
httplib2==0.21.0
idna==3.4
jws==0.1.3
oauth2client==4.1.3
Pillow==9.4.0
protobuf==4.21.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.15.0
PyJWT==2.6.0
pyparsing==3.0.9
Pyrebase==3.0.27
Pyrebase4==4.5.0
python-jwt==2.0.1
pytz==2022.6
requests==2.28.1
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12
